[fix] use the same default image for Classifier and Annotator, resolve #321
